Complete Exercise 4.21

+++ part4/controllers/blogs.js
@@ -2,7 +2,6 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
-const { error } = require('../utils/logger')
 const tokenExtractor = require('../utils/middleware')
 
 const getTokenFrom = request => {
@@ -24,6 +23,7 @@ blogsRouter.post('/',  async (request, response) => {
     if (!decodedToken.id) {
       return response.status(401).json({ error: 'token invalid' })
     }
+    console.log(decodedToken.id)
     const user = await User.findById(decodedToken.id)
     
     const blog = new Blog({
@@ -45,10 +45,19 @@ blogsRouter.post('/',  async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
     const id = request.params.id
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
 
-    const deletedBlog = await Blog.findByIdAndDelete(id)
-
+    const deletedBlog = await Blog.findById(id)
+    
     if (deletedBlog) {
+      if (deletedBlog.user.toString() !== decodedToken.id) {
+        console.log(deletedBlog.user.id.toString())
+        return response.status(401).json({error: 'invalid user'})
+      }
+      await Blog.findByIdAndDelete(id)
       response.status(204).end(); 
   } else {
       response.status(404).json({ error: 'Blog not found' });
